+++ aa-display/src/main/java/io/github/nitsuya/aa/display/xposed/hook/AndroidAutoHook.kt
@@ -8,6 +8,7 @@ import com.github.kyuubiran.ezxhelper.utils.hookBefore
 import de.robv.android.xposed.XC_MethodHook
 import de.robv.android.xposed.callbacks.XC_LoadPackage
 import io.github.nitsuya.aa.display.xposed.hook.aa.AaBtnEventHook
+import io.github.nitsuya.aa.display.xposed.hook.aa.AaClusterLyricEgressHook
 import io.github.nitsuya.aa.display.xposed.hook.aa.AaFrxRequiredAppsHook
 import io.github.nitsuya.aa.display.xposed.hook.aa.AaMediaPlaceholderHook
 import io.github.nitsuya.aa.display.xposed.hook.aa.AaNavFallbackHook
@@ -40,6 +41,7 @@ object AndroidAutoHook : BaseHook() {
             AaFrxRequiredAppsHook,
             AaNavFallbackHook,
             AaMediaPlaceholderHook,
+            AaClusterLyricEgressHook,
         ).filter { i -> i.isSupportProcess(processName) }
         if(hooks.isEmpty()) return